+++ Changelog.md
@@ -7,6 +7,12 @@ This project does _not_ adhere to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) but contrary to it
 every new version is a new major version.
 
+## 79 - 2023-08-04
+
+### Added
+
+-   Export app utility functions (like `isInstalled`) and types.
+
 ## 78 - 2023-08-03
 
 ### Changed